+++ src/renderer/menu/menu.ts
@@ -13,7 +13,7 @@ import './theme.scss';
 
 import { EventEmitter } from 'events';
 
-import * as math from './math';
+import * as math from '../math';
 import { IVec2 } from '../../common';
 import { IRenderNode } from './node';
 import { GestureDetection } from './gesture-detection';
